requirements.txt
sklearn
tensorflow-gpu
keras